% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_utiles.R
\name{char_summary}
\alias{char_summary}
\title{This function creates summary of character vector}
\usage{
char_summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{a character vector hhhhhh}

\item{na.rm}{an optional logical parameter for TRUE by default}
}
\value{
A named character with six values
}
\description{
This function creates summary of character vector
}
\examples{
\dontrun{
 y<-c("MA","IL","ME","CA")
 char_summary(y)
 }
}
