% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_utiles.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Title}
\usage{
numeric_summary(x, na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{a numeric}

\item{na.rm}{something}

\item{verbose}{is TRUE for verbose output}
}
\description{
Title
}
